Extract json encoder into base, add tests.
